fix: ES2022 Class accessor property fixes
- Codegen: Fix illegal field names for accessor methods (spaces, minus, plus, digits)
- ScriptRuntime: Preserve getter/setter for numeric property keys
- IRFactory: Fix numeric key truncation and computed property key handling
- NativeClass: Set prototypeProperty member for correct .prototype access
- Parser: Fix numeric key truncation in destructuring assignment

Fixes:
- ClassFormatError for accessor names like 'get 2.0', 'get 1E9'
- Numeric getter/setter properties lost on class prototype
- Computed property accessor not working correctly
- C.prototype returning undefined vs C['prototype'] working
- Decimal key truncation (0.1 -> 0) in various contexts

+++ rhino/src/main/java/org/mozilla/javascript/Parser.java
@@ -5239,7 +5239,8 @@ public class Parser {
                 rightElem = new Node(Token.GETPROP, createName(tempName), s);
                 extractedKeys.add(keyName);
             } else if (id instanceof NumberLiteral) {
-                Node s = createNumber((int) ((NumberLiteral) id).getNumber());
+                // Use double value for numeric keys to preserve decimals
+                Node s = createNumber(((NumberLiteral) id).getNumber());
                 rightElem = new Node(Token.GETELEM, createName(tempName), s);
             } else if (id instanceof ComputedPropertyKey) {
                 // handle computed property key: [...key]
